GCPInvoiceParser/requirements.txt
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.6.0
google-cloud-documentai==2.20.2
google-cloud-storage==1.36.2
google-crc32c==1.5.0
google-resumable-media==1.3.3
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.48.2
idna==3.6
install==1.3.5
numpy==1.26.2
packaging==23.2
pandas==2.1.3
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
simplejson==3.19.2
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
